+++ Engine/Include/Triangle.h
@@ -17,7 +17,7 @@ namespace engine
         virtual const bool init(void);
         virtual const bool init(const Vec3 & vertex_1, const Vec3 & vertex_2, const Vec3 & vertex_3);
 
-        virtual const bool render(const float dp);
+        virtual const bool render(const int dp);
         
         ~Triangle(void){}
     private:
